requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
Flask-Session==0.8.0
Flask-WTF==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.0
pillow==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2022.1
reportlab==4.3.1
requests==2.25.1
Werkzeug==2.3.7
WTForms==3.2.1
Babel==2.8.0
email-validator==2.0.0.post2
itsdangerous==2.1.2
cachelib==0.13.0
blinker==1.9.0
waitress==2.1.2
